Clean up Duration/Instant naming inconsistencies found in BCV audit
- Mark Duration-typed members @JvmSynthetic where a Java-friendly
  counterpart already exists (Airship.waitForReadyBlocking,
  AirshipConfigOptions.backgroundReportingInterval/Builder.set...,
  AutomationSchedule.interval/Builder.setInterval,
  InAppMessagingInterface.displayInterval, Banner.duration/
  DEFAULT_DURATION, LegacyInAppMessage.displayDuration,
  PushProviderBridge...setMaxCallbackWaitTime), so Java no longer sees
  an uncallable, name-mangled stub alongside the real entry point.
- Rename AirshipConfigOptions.backgroundReportingIntervalMS /
  Builder.setBackgroundReportingIntervalMS to the Ms casing used
  everywhere else in the SDK; the old MS-cased names remain as
  @Deprecated pass-throughs with no duplicate backing storage.
- Change AutomationSchedule.editGracePeriodDays from ULong? to Long?,
  since a signed Long has no realistic capacity concern for a day
  count and needs no Java-incompatible value-class handling.

+++ urbanairship-automation/src/main/java/com/urbanairship/automation/storage/AutomationStoreMigrator.kt
@@ -146,7 +146,7 @@ internal class AutomationStoreMigrator(
                     delay = getDelay(fullSchedule),
                     metadata = fullSchedule.schedule.metadata?.toJsonValue(),
                     campaigns = fullSchedule.schedule.campaigns,
-                    editGracePeriodDays = TimeUnit.MILLISECONDS.toDays(fullSchedule.schedule.editGracePeriod).toULong(),
+                    editGracePeriodDays = TimeUnit.MILLISECONDS.toDays(fullSchedule.schedule.editGracePeriod),
                     productId = fullSchedule.schedule.productId,
                     frequencyConstraintIds = fullSchedule.schedule.frequencyConstraintIds?.ifEmpty { null },
                     messageType = fullSchedule.schedule.messageType,
